+++ decidim-admin/app/controllers/decidim/admin/dashboard_controller.rb
@@ -5,7 +5,9 @@ module Decidim
     # Controller that shows a simple dashboard.
     #
     class DashboardController < Decidim::Admin::ApplicationController
-      helper_method :latest_action_logs, :users_counter
+      helper_method :latest_action_logs
+      helper_method :users_counter
+      helper_method :metrics_presenter
 
       def show
         enforce_permission_to :read, :admin_dashboard
@@ -19,7 +21,14 @@ module Decidim
                                 .includes(:participatory_space, :user, :resource, :component, :version)
                                 .for_admin
                                 .order(created_at: :desc)
-                                .first(20)
+                                .first(5)
+      end
+
+      def metrics_presenter
+        @metrics_presenter ||= Decidim::Admin::DashboardMetricChartsPresenter.new(
+          summary: true,
+          organization: current_organization
+        )
       end
 
       def users_counter
@@ -27,7 +36,7 @@ module Decidim
         last_week = Time.zone.today.prev_week
         last_month = Time.zone.today.prev_month
 
-        @result = {
+        {
           total_admins_last_24: users_count(last_day, true),
           total_admins_last_week: users_count(last_week, true),
           total_admins_last_month: users_count(last_month, true),